+++ asf_search/ASFProduct.py
@@ -18,6 +18,7 @@ class ASFProduct:
         self.properties = translated['properties']
         self.geometry = translated['geometry']
         self.baseline = translated['baseline']
+        self.searchOptions = opts
 
     def __str__(self):
         return json.dumps(self.geojson(), indent=2, sort_keys=True)
@@ -35,12 +36,15 @@ class ASFProduct:
 
         :param path: The directory into which this product should be downloaded.
         :param filename: Optional filename to use instead of the original filename of this product.
-        :param session: The session to use, in most cases should be authenticated beforehand
+        :param session: The session to use, defaults to the one used to find the results.
 
         :return: None
         """
         if filename is None:
             filename = self.properties['fileName']
+        
+        if session is None and self.searchOptions is not None:
+            session = self.searchOptions.session
 
         download_url(url=self.properties['url'], path=path, filename=filename, session=session)
 
@@ -56,7 +60,8 @@ class ASFProduct:
         :return: ASFSearchResults containing the stack, with the addition of baseline values (temporal, perpendicular) attached to each ASFProduct.
         """
         from .search.baseline_search import stack_from_product
-
+        # *this* opts, probably isn't the same as the one used to do the search.
+        # Don't default to self.searchOptions here
         return stack_from_product(self, opts=opts)
 
     def get_stack_opts(self) -> ASFSearchOptions: